# api.yaml
openapi: 3.0.0
info:
  title: Track Selection Service API
  description: API для сервиса выбора трека обучения
  version: 1.0.0
  contact:
    name: Support
    email: support@track-selection.com
servers:
  - url: http://localhost:8080/api
    description: Local development server
  - url: https://api.track-selection.com/api
    description: Production server
...
paths:
...
  /student/recommendations:
    get:
      tags:
        - Recommendations
      summary: Получить рекомендации треков
      description: |
        Возвращает список треков, отсортированных по степени соответствия предпочтениям студента (метод PROMETHEE).
        
        **Требования для получения рекомендаций:**
        - Профессиональные цели (не пустой массив)
        - Все академические оценки (от 2 до 5)
        - Хотя бы один навык > 0
        - Стиль обучения (1, 2 или 3)
      operationId: getRecommendations
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecommendationsResponse'
        '400':
          description: Профиль не полностью заполнен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: "PROFILE_NOT_COMPLETE"
                  message: "Profile is not complete enough for recommendations"
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ==================== TRACK SELECTION ====================
  /student/select-track:
    post:
      tags:
        - Selection
      summary: Выбрать трек
      description: Студент выбирает трек для изучения
      operationId: selectTrack
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - track_id
              properties:
                track_id:
                  type: string
                  format: uuid
                  description: ID выбранного трека
      responses:
        '200':
          description: Трек успешно выбран
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "track selected"
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Студент или трек не найдены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /student/selected-tracks:
    get:
      tags:
        - Selection
      summary: Получить выбранные треки
      description: Возвращает список треков, выбранных студентом
      operationId: getSelectedTracks
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                type: object
                properties:
                  tracks:
                    type: array
                    items:
                      $ref: '#/components/schemas/Track'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /student/unselect-track/{id}:
    delete:
      tags:
        - Selection
      summary: Отменить выбор трека
      description: Удаляет выбранный трек из списка студента
      operationId: unselectTrack
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID трека для отмены выбора
      responses:
        '200':
          description: Выбор успешно отменен
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "track unselected"
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Выбор не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
...
    Course:
      type: object
      description: Информация о курсе в семестре
      properties:
        name:
          type: string
          description: Название курса
          example: "Введение в программирование"
        description:
          type: string
          description: Описание курса
          example: "Основы Go, синтаксис, переменные, функции"
        is_elective:
          type: boolean
          description: "false - обязательный курс, true - курс по выбору"
          example: false
        options:
          type: array
          description: "Варианты выбора (если is_elective = true)"
          items:
            type: string
          example: ["A2", "B1", "B2"]
    Semester:
      type: object
      description: Семестр с курсами
      properties:
        number:
          type: integer
          description: Номер семестра
          example: 1
        courses:
          type: array
          description: Список курсов в семестре
          items:
            $ref: '#/components/schemas/Course'
    YearTrack:
      type: object
      description: Трек на один год (для типа single)
      properties:
        name:
          type: string
          description: Название трека на год
        description:
          type: string
          description: Описание
        semesters:
          type: array
          description: Семестры в этом году
          items:
            $ref: '#/components/schemas/Semester'
    YearBranch:
      type: object
      description: Ветка выбора (для типа branching)
      properties:
        name:
          type: string
          description: Название подтрека
        description:
          type: string
          description: Описание
        semesters:
          type: array
          description: Семестры в этом подтреке
          items:
            $ref: '#/components/schemas/Semester'
    YearPlan:
      type: object
      description: План на один учебный год
      properties:
        year:
          type: integer
          description: Номер года обучения (3,4,5,6,7,8 семестр)
          example: 3
        type:
          type: string
          enum: [single, branching]
          description: "single - один трек, branching - выбор подтрека"
        track:
          $ref: '#/components/schemas/YearTrack'
        branches:
          type: array
          description: Список подтреков для выбора (при type=branching)
          items:
            $ref: '#/components/schemas/YearBranch'
    Curriculum:
      type: object
      description: Учебный план, разбитый по годам с возможностью выбора подтреков
      properties:
        years:
          type: array
          description: Массив учебных лет
          items:
            $ref: '#/components/schemas/YearPlan'
    Requirement:
      type: object
      properties:
        subject:
          type: string
          description: Название предмета
          example: "programming"
        min_grade:
          type: integer
          minimum: 2
          maximum: 5
          description: Минимальная оценка для поступления
          example: 4
    Track:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        curriculum:
          $ref: '#/components/schemas/Curriculum'
        requirements:
          type: array
          items:
            $ref: '#/components/schemas/Requirement'
        teachers:
          type: array
          items:
            type: string
        difficulty:
          type: integer
          minimum: 1
          maximum: 5
          description: "Сложность трека (1 - легкий, 5 - очень сложный)"
          example: 3
        type:
          type: integer
          description: "Тип трека для фронтенда (1 - технический, 2 - аналитический и т.д.)"
          example: 1
        employment_prospects:
          type: integer
          minimum: 1
          maximum: 10
        alumni_reviews:
          type: integer
          minimum: 1
          maximum: 10
        web_link:
          type: string
        has_certificates:
          type: integer
          enum: [0, 1]
        learning_style:
          type: integer
          enum: [1, 2, 3]
        desired_tech_skills:
          type: integer
          minimum: 1
          maximum: 10
        desired_math_skills:
          type: integer
          minimum: 1
          maximum: 10
        desired_soft_skills:
          type: integer
          minimum: 1
          maximum: 10
        professional_goals:
          type: array
          items:
            type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    TrackScore:
      type: object
      properties:
        track_id:
          type: string
          format: uuid
        track_name:
          type: string
        score:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: "Итоговая оценка трека (0-1)"
        rank:
          type: integer
          description: "Место в рейтинге"
        criteria_scores:
          type: object
          additionalProperties:
            type: number
          description: "Оценки по каждому критерию"
    RecommendationsResponse:
      type: object
      properties:
        recommendations:
          type: array
          items:
            $ref: '#/components/schemas/TrackScore'
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []